banking-sync/requirements.txt
flask>=3.0.0
requests>=2.31.0
apscheduler>=3.10.4
pyyaml>=6.0.1
PyJWT>=2.8.0
cryptography>=42.0.0
